Add local storage

+++ src/components/pages/Projects.js
@@ -7,6 +7,7 @@ import LinkButton from "../layout/Buttons/LinkButton";
 import ProjectCard from "../project/ProjectCard";
 import Loading from "../layout/Loading";
 
+import internaldb from '../../internaldb.json'
 
 import styles from './pagesCss/Projects.module.css'
 
@@ -15,23 +16,43 @@ export default function Projects() {
     const [projects, setProjects] = useState([])
     const [projectMessage, setProjectsMessage] = useState('')
 
-    
+
+    function noProjectsAtInitial() {
+        if (!localStorage.hasOwnProperty('projects')) {
+            localStorage.setItem('projects', JSON.stringify(internaldb.projects))
+        }
+    }
+
+
     useEffect(() => {
         setTimeout(() => {
-            fetch('http://localhost:5000/projects', {
-                method: 'GET',
-                headers: {
-                    'content-Type': 'application/json'
-                }
-            })
-            .then((resp) => resp.json())
-            .then((data) => {
-                setProjects(data)
+
+            if (window.location.hostname === 'localhost' || window.location.hostname === '127.0.0.1') {
+
+                fetch('http://localhost:5000/projects', {
+                    method: 'GET',
+                    headers: {
+                        'content-Type': 'application/json'
+                    }
+                })
+                .then((resp) => resp.json())
+                .then((data) => {
+                    setProjects(data)
+                    setRemoveLoading(true)
+                })
+                .catch((err) => console.log(err))
+                
+            }else{
+                // USING LOCAL STORAGE
+                noProjectsAtInitial()
+                const localProjects = localStorage.getItem('projects')
+                setProjects(JSON.parse(localProjects))
                 setRemoveLoading(true)
-            })
-            .catch((err) => console.log(err))
+            }
+
         }, 500)
     }, [])
+
 
     const location = useLocation()
     let message = ''
@@ -39,20 +60,34 @@ export default function Projects() {
         message = location.state.message
     }
 
+
     function removeProject(id) {
-        fetch(`http://localhost:5000/projects/${id}`, {
-            method: 'DELETE',
-            headers: {
-                'content-Type': 'application/json'
-            }
-        })
-        .then((resp) => resp.json())
-        .then(() => {
-            setProjects(projects.filter((project) => project.id !== id))
-            setProjectsMessage('Projeto removido com sucesso!')
-        })
-        .catch((err) => console.log(err))
+
+        if (window.location.hostname === 'localhostd' || window.location.hostname === '127.0.0.1') {
+
+            fetch(`http://localhost:5000/projects/${id}`, {
+                method: 'DELETE',
+                headers: {
+                    'content-Type': 'application/json'
+                }
+            })
+            .then((resp) => resp.json())
+            .then(() => {
+                setProjects(projects.filter((project) => project.id !== id))
+                setProjectsMessage('Projeto removido com sucesso!')
+            })
+            .catch((err) => console.log(err))
+
+        }else{
+            // USING LOCAL STORAGE
+            const localProjects = JSON.parse(localStorage.getItem('projects'))
+            const updatedProject = localProjects.filter(project => project.id !== id)
+            let updateJsonString = JSON.stringify(updatedProject)
+            localStorage.setItem('projects', updateJsonString)
+            setProjects(JSON.parse(localStorage.getItem('projects')))
+        }
     }
+
 
     return (
         <div className={styles.project_container}>
